Move useCreateExistingCards to own file
